Add files via upload
Function 1 (20 pts):
Name of the function: print_square
Return type: void
Input parameters: 1 integer (side)
This function prints a square of asterisks (*) with the side length provided by the user. The pattern is
printed as follows:
****
****
****
****
Use for loop to print the pattern
Function 2 (20 pts):
Name of the function: print_right_triangle
Return type: void
Input parameters: 1 integer (height)
This function prints a right-angled triangle of asterisks (*) with the height provided by the user. The
pattern is printed as follows:
*
**
***
****
Use for loop to print the pattern
Function 3 (20 pts):
Name of the function: print_number_pattern
Return type: void
Input parameters: 1 integer (rows)
This function prints a pattern of numbers. The number of rows is provided by the user. The pattern is
printed as follows:
1
22
333
4444
55555
Use for loop to print the pattern

Main Function (40 pts):
Name of the function: main
Return type: int
Input parameters: none
Declare a variable to store the type of the pattern (char).
Ask the user to enter the type of the pattern. The user can enter 's' for square, 't' for rightangled triangle and 'n' for number pattern.
If the user enters a character other than 's', 't' or 'n', print "Invalid pattern type" and ask
the user to enter the type of the pattern again. Repeat this process until the user enters a
valid pattern type.
If the user enters 's', ask the user to enter the side length of the square. Then, call the
print_square function to print the square.
If the user enters 't', ask the user to enter the height of the right-angled triangle. Then, call the
print_right_triangle function to print the right-angled triangle.
If the user enters 'n', ask the user to enter the number of rows in the number pattern. Then, call
the print_number_pattern function to print the number pattern.
Use while loop to ask the user to enter the number of rows and the type of the pattern.